Also more workspace search into find/
